make the port as ushort

+++ MLS.Agent/CommandLine/StartupOptions.cs
@@ -43,7 +43,7 @@ namespace MLS.Agent.CommandLine
             string package = null,
             string packageVersion = null,
             ParseResult parseResult = null,
-            int? port = null)
+            ushort? port = null)
         {
             _parseResult = parseResult;
             LogPath = logPath;
@@ -100,6 +100,6 @@ namespace MLS.Agent.CommandLine
         public string Package { get; }
 
         public string PackageVersion { get; }
-        public int? Port { get; }
+        public ushort? Port { get; }
     }
 }
